Boot script: Switches to virtio-blk, moves installer disk options
Using virtio-blk seems to be supported natively in modern macOS versions. This should be faster and also allows us to mark the installer image as read-only, preventing any modification by the guest. Additionally, this moves the installer image related command line components out of the main block of the Qemu command line, so that it's easy to comment them out once installation has completed, or indeed to wrap them in a conditional block.